GeradorContratos/requirements.txt
lxml==4.9.3
python-docx==1.1.0
typing_extensions==4.8.0
